+++ NEWS.md
@@ -1,4 +1,21 @@
-# arkdb 0.0.3
+
+# arkdb 0.0.4
+
+- `unark()` will strip out non-compliant characters in table names by default.
+- `unark()` gains the optional argument `tablenames`, allowing the user to
+   specify the corresponding table names manually, rather than enforcing
+   they correspond with the incoming file names. 
+   [#18](https://github.com/ropensci/arkdb/issues/18)
+-  `unark()` gains the argument `encoding`, allowing users to directly set
+   the encoding of incoming files.  Previously this could only be set by
+   setting `options(encoding)`, which will still work as well. See
+  `FAO.R` example in `examples` for an illustration.  
+- `unark()` will now attempt to guess which streaming parser to use 
+   (e.g `csv` or `tsv`) based on the file extension pattern, rather than
+   defaulting to a `tsv` parser.  (`ark()` still defaults to exporting in
+   the more portable `tsv` format).
+
+# arkdb 0.0.3 2018-09-11
 
 * Remove dependency on utils::askYesNo for backward compatibility, [#17](https://github.com/ropensci/arkdb/issues/17)